+++ src/ui/src/highlightersetedit.cpp
@@ -94,7 +94,7 @@ HighlighterSetEdit::HighlighterSetEdit( QWidget* parent )
              &HighlighterSetEdit::updateHighlighterProperties );
 
     dispatchToMainThread( [ this ] {
-        IconLoader iconLoader( this );
+        IconLoader iconLoader;
 
         addHighlighterButton->setIcon( iconLoader.load( "icons8-plus-16" ) );
         removeHighlighterButton->setIcon( iconLoader.load( "icons8-minus-16" ) );