+++ src/ui/src/tabbedcrawlerwidget.cpp
@@ -34,11 +34,11 @@
 #include "crawlerwidget.h"
 
 #include "clipboard.h"
-#include "configuration.h"
 #include "dispatch_to.h"
 #include "iconloader.h"
 #include "log.h"
 #include "openfilehelper.h"
+#include "tabbarstyle.h"
 #include "tabgroupinfo.h"
 #include "tabnamemapping.h"
 #include "theme.h"
@@ -61,55 +61,7 @@ TabbedCrawlerWidget::TabbedCrawlerWidget()
     , newdata_icon_( ":/images/newdata_icon.png" )
     , newfiltered_icon_( ":/images/newfiltered_icon.png" )
 {
-
-    const auto& config = Configuration::get();
-    const bool isDark = ( config.style() == Theme::DarkKey );
-
-    QString tabStyle = QStringLiteral( "QTabBar::tab { height: 28px; }" );
-
-    QString tabCloseButtonStyle = " QTabBar::close-button { %1 }";
-
-    QString backgroundImage;
-    QString backgroundHoverImage;
-
-    if ( isDark ) {
-        backgroundImage = ":/images/icons8-close-window-16_inverse.png";
-        backgroundHoverImage = ":/images/icons8-close-window-hover-16_inverse.png";
-    }
-
-#if defined( Q_OS_MAC )
-    // work around Qt macOS bug missing tab close icons
-    // see: https://bugreports.qt.io/browse/QTBUG-61092
-    if ( !isDark ) {
-        backgroundImage
-            = ":/qt-project.org/styles/commonstyle/images/standardbutton-closetab-16.png";
-        backgroundHoverImage
-            = ":/qt-project.org/styles/commonstyle/images/standardbutton-closetab-hover-16.png";
-    }
-#elif defined( Q_OS_WIN )
-    if ( config.style() == Theme::LightKey ) {
-        backgroundImage = ":/images/icons8-close-window-16.png";
-        backgroundHoverImage = ":/images/icons8-close-window-hover-16.png";
-    }
-#endif
-
-    if ( !backgroundImage.isEmpty() ) {
-        const QString backgroundImageTemplate = " image: url(%1);";
-        QString tabCloseButtonHoverStyle
-            = isDark ? " QTabBar::close-button:hover { %1 background-color: #C42B1C;"
-                       " border-radius: 3px; }"
-                     : " QTabBar::close-button:hover { %1 }";
-        backgroundImage = backgroundImageTemplate.arg( backgroundImage );
-        backgroundHoverImage = backgroundImageTemplate.arg( backgroundHoverImage );
-        tabCloseButtonHoverStyle = tabCloseButtonHoverStyle.arg( backgroundHoverImage );
-        tabCloseButtonStyle = tabCloseButtonStyle.arg( backgroundImage );
-        tabCloseButtonStyle.append( tabCloseButtonHoverStyle );
-    }
-    else {
-        tabCloseButtonStyle = tabCloseButtonStyle.arg( "" );
-    }
-
-    myTabBar_.setStyleSheet( tabStyle + tabCloseButtonStyle );
+    applyTheme();
 
     setTabBar( &myTabBar_ );
     myTabBar_.hide();
@@ -121,9 +73,14 @@ TabbedCrawlerWidget::TabbedCrawlerWidget()
     dispatchToMainThread( [ this ] { loadIcons(); } );
 }
 
+void TabbedCrawlerWidget::applyTheme()
+{
+    myTabBar_.setStyleSheet( closableTabBarStyleSheet( Theme::active() ) );
+}
+
 void TabbedCrawlerWidget::loadIcons()
 {
-    IconLoader iconLoader{ this };
+    IconLoader iconLoader;
     olddata_icon_ = iconLoader.load( "olddata_icon" );
     for ( int tab = 0; tab < count(); ++tab ) {
         updateIcon( tab );